add extra details to hapi swagger info
