+++ src/api/index.js
@@ -11,23 +11,7 @@ import { secureContext } from './common/helpers/secure-context/index.js'
 import { pulse } from './common/helpers/pulse.js'
 import { requestTracing } from './common/helpers/request-tracing.js'
 import { setupProxy } from './common/helpers/proxy/setup-proxy.js'
-
-const hapiSwaggerOptions = {
-  info: {
-    title: 'FCP SFD Data API',
-    version: '0.0.1'
-  },
-  OAS: 'v3.0',
-  documentationPath: '/documentation',
-  jsonPath: '/documentation.json',
-  cors: false,
-  servers: [
-    {
-      url: 'http://localhost:3000',
-      description: 'local server'
-    }
-  ]
-}
+import { hapiSwaggerOptions } from './common/helpers/hapi-swagger-options.js'
 
 const createServer = async () => {
   setupProxy()
